Revert move of migratePersistentActiveBakers.
