+++ concordium-consensus/src/Concordium/GlobalState/Persistent/BlockState.hs
@@ -202,45 +202,6 @@ migrateSeedStateV1Trivial SeedStateV1{..} =
     newNonce = H.hash $ "Regenesis" <> encode ss1UpdatedNonce
 
 -- | See documentation of @migratePersistentBlockState@.
-migratePersistentActiveBakers ::
-    forall oldpv pv t m.
-    ( IsProtocolVersion oldpv,
-      IsProtocolVersion pv,
-      SupportMigration m t,
-      SupportsPersistentAccount pv (t m)
-    ) =>
-    StateMigrationParameters oldpv pv ->
-    -- | Already migrated accounts.
-    Accounts.Accounts pv ->
-    PersistentActiveBakers (AccountVersionFor oldpv) ->
-    t m (PersistentActiveBakers (AccountVersionFor pv))
-migratePersistentActiveBakers migration accounts PersistentActiveBakers{..} = do
-    newActiveBakers <- Trie.migrateTrieN True (migratePersistentActiveDelegators migration) _activeBakers
-    newAggregationKeys <- Trie.migrateTrieN True return _aggregationKeys
-    newPassiveDelegators <- migratePersistentActiveDelegators migration _passiveDelegators
-    bakerIds <- Trie.keysAsc newActiveBakers
-    totalStakedAmount <-
-        foldM
-            ( \acc (BakerId aid) ->
-                Accounts.indexedAccount aid accounts >>= \case
-                    Nothing -> error "Baker account does not exist."
-                    Just pa ->
-                        accountBakerStakeAmount pa >>= \case
-                            Nothing -> error "Baker account not a baker."
-                            Just amt -> return $! (acc + amt)
-            )
-            0
-            bakerIds
-    let newTotalActiveCapital = migrateTotalActiveCapital migration totalStakedAmount _totalActiveCapital
-    return
-        PersistentActiveBakers
-            { _activeBakers = newActiveBakers,
-              _aggregationKeys = newAggregationKeys,
-              _passiveDelegators = newPassiveDelegators,
-              _totalActiveCapital = newTotalActiveCapital
-            }
-
--- | See documentation of @migratePersistentBlockState@.
 --
 --  Migrate the birk parameters assuming accounts have already been migrated.
 migratePersistentBirkParameters ::
